Introduction to Git 12
The twelfth git practice, a very interesting topic, ignoring files, files and directories can be ignored for various reasons, for this we create a .gitignore directory in the root of our project and add everything that needs to be ignored there. If there is a committed file, to return it from the repository, but not delete it in the home directory, use the command: git rm is --cached [namefile], and only after that the file is added to .gitignore.