Introduction to Git 13
Last Git practice on the stash. Stash hides our files for commit, let's say a situation has arisen, we are working on a project and we urgently received a task that is not related to the previous task, so as not to confuse everything, we add our changes to the index and hide them using the git stash. After that, we can see that the files are gone from git status and we can commit other files, after which we can get our files back from the stack with git stash pop.